requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
torch==2.1.1
moviepy==1.0.3
flask==2.3.3 